feat: start applying MCV architecture
