+++ index.html
@@ -11,7 +11,7 @@
         integrity="sha256-WBkoXOwTeyKclOHuWtc+i2uENFpDZ9YPdf5Hf+D7ewM=" crossorigin=""></script>
 
 
-    <link rel="stylesheet" href="style/style.css">
+    <link rel="stylesheet" href="src/style/style.css">
 
     <script src="https://kit.fontawesome.com/ef4acaea7b.js" crossorigin="anonymous"></script>
     <link rel="preconnect" href="https://fonts.googleapis.com">
@@ -23,7 +23,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Akaya+Kanadaka&display=swap" rel="stylesheet">
 
-    <script type="module" src="main.js"></script>
+    <script type="module" src="src/js/controller.js"></script>
 
     <title>PtaszekTracker</title>
 </head>
@@ -129,7 +129,7 @@
         <p class="birds-list__title">Your observations: </p>
 
         <div class="bird-list__container"> -->
-            <!-- <div class="birds-list__result">
+    <!-- <div class="birds-list__result">
                 <div class="birds-list__result--icon"><i class="fa-solid fa-plus" style="color: #418900;"></i></div>
                 <div class="birds-list__heading">
                     <img src="img/animal-s.jpg" alt="Photo of the bird">
