--- main.js
@@ -1,7 +1,0 @@
-import './modules/nav.js'
-import './modules/map.js'
-import './modules/search.js'
-
-
-let date = new Date()
-export let id = Math.random() + ''
