+++ package.json
@@ -2,7 +2,7 @@
 	"name": "ptaszek-tracker",
 	"version": "1.0.0",
 	"description": "",
-	"main": "main.js",
+	"main": "controller.js",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
 		"start": "parcel index.html",
